backend/pyproject.toml
[project]
name = "gpt-battle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "psycopg[binary]>=3.2.1",
    "openai>=1.99.3",
    "loguru>=0.7.3",
    "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "modal>=1.1.1",
    "fire>=0.7.0",
]
